backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
httpx==0.27.0
pydantic==2.6.4
pydantic-settings==2.2.1
SQLAlchemy==2.0.29
aiosqlite==0.20.0
python-jose==3.3.0
cachetools==5.3.3
apscheduler==3.10.4
pytest==8.1.1
pytest-asyncio==0.23.5
groq==0.9.0
google-generativeai==0.8.3
